# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Cloudfront",
    "description": "Cloudfront api endpoints",
    "version": "1.0.0"
  },
  "host": "localhost:8006",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/facility/": {
      "post": {
        "tags": [
          "facility"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "facilityName": {
                  "example": "any"
                },
                "facilityId": {
                  "example": "any"
                },
                "assessment": {
                  "example": "any"
                },
                "realName": {
                  "example": "any"
                },
                "starRating": {
                  "example": "any"
                },
                "distance": {
                  "example": "any"
                },
                "monthlyCost": {
                  "example": "any"
                },
                "priceCategory": {
                  "example": "any"
                },
                "phoneNumber": {
                  "example": "any"
                },
                "pros": {
                  "example": "any"
                },
                "cons": {
                  "example": "any"
                },
                "recommendationReason": {
                  "example": "any"
                },
                "mapUrl": {
                  "example": "any"
                },
                "score": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
...
    "/booking/": {
      "post": {
        "tags": [
          "booking"
        ],
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "facility": {
                  "example": "any"
                },
                "type": {
                  "example": "any"
                },
                "preferredDays": {
                  "example": "any"
                },
                "preferredTimes": {
                  "example": "any"
                },
                "notes": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
...
  }
}